fix layout JS state management
Previous logic would set "jsEnabled" state to true if JS existed
on the initial page load, but on subsequent in-browser page reloads,
it wouldn't fire any more.

New logic runs every time the component is re-mounted, which means
it runs on the initial pageload and then on all rendenders thereafter.