create alternate urls for menu state in static export
this is super cool, actually.
In our config file, we're exporting two versions of each page:
- one with the closed menu
- one with the open menu

Additionally, we're looking on the server for any of these
`/m/` prefixed urls so that we can open the menu.

This is kind of a hack in the sense that we're using url paths
instead of query parameters, but it's also not a hack in that
that's just a convention.

Anyway, this is great.