add next-import-images package
This package lets us link images from a relative path in our project,
and does some compression to them.

It's not a panacea, but it's pretty good.

Otherwise, we would have to have another package run through some
public/ folder and generate a ton of smaller images.

This is an optimization we can (and should) still do, but for now
this is a pretty good start.